♻️ Firebase 관련 코드 구조 개선 및 모듈화
- auth.ts: 인증 관련 함수 분리 (회원가입, 로그인, 로그아웃, 비밀번호 재설정)
- firebaseConfig.ts: Firebase 초기화 및 기본 설정 분리
- firestore.ts: Firestore 데이터베이스 관련 함수 모듈화
- storage.ts: Firebase Storage 관련 함수 분리

🔧 각 기능별로 파일 분리로 코드 가독성 및 유지보수성 향상
🔒 인증, 데이터베이스, 스토리지 기능의 명확한 분리로 보안성 강화
📦 모듈화를 통한 코드 재사용성 증가 및 의존성 관리 개선
