+++ bagpack/frontend/src/lib/firebase/storage.ts
@@ -1,0 +1,2 @@
+import { ref, uploadBytes, getDownloadURL } from "firebase/storage";
+import { storage } from "./firebaseConfig";
